+++ src/reader/core.py
@@ -606,7 +606,13 @@ class Reader:
         url = _feed_argument(feed)
         self._storage.set_feed_updates_enabled(url, False)
 
-    def update_feeds(self, new_only: bool = False, workers: int = 1) -> None:
+    def update_feeds(
+        self,
+        new_only: Union[bool, MissingType] = MISSING,
+        workers: int = 1,
+        *,
+        new: Union[Optional[bool], MissingType] = MISSING,
+    ) -> None:
         """Update all the feeds that have updates enabled.
 
         Silently skip feeds that raise :exc:`ParseError`.
@@ -614,8 +620,17 @@ class Reader:
         Roughly equivalent to ``for _ in reader.update_feed_iter(...): pass``.
 
         Args:
-            new_only (bool): Only update feeds that have never been updated.
+            new_only (bool):
+                Only update feeds that have never been updated.
+                Defaults to False.
+
+                .. deprecated:: 1.19
+                    Use ``new`` instead.
+
             workers (int): Number of threads to use when getting the feeds.
+            new (bool or None):
+                Only update feeds that have never been updated
+                / have been updated before. Defaults to None.
 
         Raises:
             StorageError
@@ -635,8 +650,13 @@ class Reader:
             Previously, entries would be updated only if the
             entry :attr:`~Entry.updated` was *newer* than the stored one.
 
+        .. deprecated:: 1.19
+            The ``new_only`` argument
+            (will be removed in *reader* 2.0);
+            use ``new`` instead.
+
         """
-        for url, value in self.update_feeds_iter(new_only, workers):
+        for url, value in self.update_feeds_iter(new_only, workers, new=new):
             if isinstance(value, ParseError):
                 log.exception(
                     "update feed %r: error while getting/parsing feed, "
@@ -650,13 +670,26 @@ class Reader:
             assert not isinstance(value, Exception), value
 
     def update_feeds_iter(
-        self, new_only: bool = False, workers: int = 1
+        self,
+        new_only: Union[bool, MissingType] = MISSING,
+        workers: int = 1,
+        *,
+        new: Union[Optional[bool], MissingType] = MISSING,
     ) -> Iterable[UpdateResult]:
         """Update all the feeds that have updates enabled.
 
         Args:
-            new_only (bool): Only update feeds that have never been updated.
+            new_only (bool):
+                Only update feeds that have never been updated.
+                Defaults to False.
+
+                .. deprecated:: 1.19
+                    Use ``new`` instead.
+
             workers (int): Number of threads to use when getting the feeds.
+            new (bool or None):
+                Only update feeds that have never been updated
+                / have been updated before. Defaults to None.
 
         Yields:
             :class:`UpdateResult`:
@@ -680,16 +713,34 @@ class Reader:
             Update entries whenever their content changes.
             See :meth:`~Reader.update_feeds` for details.
 
+        .. deprecated:: 1.19
+            The ``new_only`` argument
+            (will be removed in *reader* 2.0);
+            use ``new`` instead.
+
         """
         if workers < 1:
             raise ValueError("workers must be a positive integer")
 
+        if new is MISSING and new_only is MISSING:
+            new_final = None
+        elif new is MISSING and new_only is not MISSING:
+            new_final = True if new_only else None
+            warnings.warn(
+                "new_only is deprecated and will be removed in reader 2.0. "
+                "Use new instead.",
+                DeprecationWarning,
+            )
+        elif new is not MISSING and new_only is MISSING:
+            assert not isinstance(new, MissingType)  # mypy pleasing
+            new_final = new
+        else:
+            raise TypeError("new and new_only are mutually exclusive")
+
         make_map = nullcontext(builtins.map) if workers == 1 else make_pool_map(workers)
 
-        new = True if new_only else None
-
         with make_map as map:
-            results = self._update_feeds(new=new, map=map)
+            results = self._update_feeds(new=new_final, map=map)
 
             for url, value in results:
                 if isinstance(value, FeedNotFoundError):
@@ -1175,7 +1226,7 @@ class Reader:
                     f"get_feed_metadata() takes 1 positional arguments, but {len(args) + 1} were given"
                 )
             warnings.warn(
-                "the get_feed_metadata(feed, key[, default]) -> value "
+                "The get_feed_metadata(feed, key[, default]) -> value "
                 "version of get_feed_metadata() is deprecated "
                 "and will be removed in reader 2.0. "
                 "Use get_feed_metadata_item() instead.",